+++ src/undate/__init__.py
@@ -1,14 +1,17 @@
-__version__ = "0.6.1"
+__version__ = "0.7.0"
 
+# this sort order is important to avoid circular imports
+
+# ruff: noqa: I001
 from undate.date import DatePrecision, UnDelta
-from undate.undate import Undate, Calendar
+from undate.undate import Calendar, Undate
 from undate.interval import UndateInterval
 
 __all__ = [
-    "Undate",
-    "UndateInterval",
     "Calendar",
     "DatePrecision",
     "UnDelta",
+    "Undate",
+    "UndateInterval",
     "__version__",
 ]